requirements.txt
langchain
openai
tiktoken
faiss-cpu
pypdf
streamlit
streamlit-extras